fix: allow for multiple s3 object selections
This PR leverages that and related s3 improvements for
1) ml/codeflare/training/bert: store tensorboard events in s3
2) ml/codeflare/training/glue: choose glue-data and model

This PR also fixes aggressive validation checks in ml/ray/hcks/openshift/uid-range. We don't need to bother with potentially expensive kubectl operations if KUBE_CONTEXT and KUBE_NS are not defined. That guidebook was also not using KUBE_CONTEXT In its kubectl commands.