fix: change `--context ${KUBE_CONTEXT}` to `${KUBE_CONTEXT_ARG}`
And the same for `-n ${KUBE_NS}` to `${KUBE_NS_ARG}`.

This allows us to run in situations where we do not want to set any `--context` or `-n` options. For example, when running inside a kubernetes cluster, `kubectl` is smart enough to pick up this information automatically. I don't see any way to pass some kind of fake context name in-cluster. For namespace... this *is* knowable, even in-cluster, but I figured we should make this change uniformly here.