fix: client-side of in-cluster log aggregator may fail to stream certain content
If the client-side starts streaming before the server-side has any data, the tail -f's will exit and never recover.

This PR introduces a script `wait-for-and-tailf` which relies on `inotifywait` (part of `inotify-tools` for ubuntu) to wait for a file to exist (using push notifications from inotify), and then start tailing.

TODO: how can we get this utility into the store?