fix: allow splicing in of namespace in ray helm chart
Oof, mcad currently has a bug that forces us to state explicitly the target namespace in all of our resources. When running in-cluster, it may be difficult to intuit this name, especially under test scenarios (where, in order to use a fixed namespace for the guidebook profile (the answer to the "namespace" question).

This PR allows clients to set a KUBE_NS_FOR_REAL env var that will be used just for this scenario. Once mcad fixes its bug, i think we can back this out.