fix: improve git compatibility with sparse checkout
ugh, it looks like git had a horrible bug in its git checkout --sparse when you use a url, and of course broken versions of git seem to be fairly common  — e.g. a broken version was installed in my ubuntu-on-WSL2 by default.

we were using git clone --sparse https://somethingsomething, which git 2.25 does not like. This PR switches us to a form that seems to work with 2.25 and later versions as well (2.37.x tested).

all of our automated tests apparently were using a version without this git bug.