feat: submit ray job in helm chart
This also switches ray head and worker from Deployment to Job. This is done so that we can exit the head and worker pods ASAP upon job completion. Deployments only allow for an Always restart policy. Jobs, on the other hand, allow for `OnFailure`, which I think is what we want.